Now we're doing non-greedy quote matching (fixes issue 3).
First, we look for shortest-possible open/close single-quote pairs and replace
them with the appropriate Unicode. Then, we replace any single quotes left
with apostrophes. This takes care of most abbreviations and edge cases.

New issue
Summary: test case "testPrettifiesQuotesAtStartOfLine" is broken
Labels: Component-Logic
The test for prettifying quotes at the start of the line is broken: it assumes
greedy replacement, so this revision causes it to fail.

Update issue 14
The test case specified now passes, but the issue is still outstanding, as
evidenced by the failure of testMatchesQuotesAcrossTags. This will probably
require some extra fiddling with prettifyHtml and prettifyStr alike.