Failing test for contiguous quote pairs.
New issue
Summary: contiguous single quote pairs are not properly prettified.
When contiguous single quote pairs, e.g. "'a' 'b' 'c'", are prettified, every
other one ("'b'" in this case) doesn't get matched as a quote pair. The
problem is that the regex eats the space character in the previous match, so
the rule doesn't get triggered if it's only run once.