Bookmarklet is now hosted on Google AppEngine.
New issue
Summary: There's no way to be sure that I'm not evil
By having the bookmarklet grab a copy of prettify-comp.js from Google
Code, we could ensure that I wasn't doing anything malicious: any
modification made to the script would show up in the revision log. Now,
there's nothing stopping me from redirecting the script to evil.js which
harvests your credentials and sends profane emails to your boss.
Granted, I promise not to, but if you're paranoid, that might not bew
good enough.

Ideally, we could verify a hash or signature on the script file to
ensure it hadn't been tampered with. This is hard to do in a
bookmarklet, though, especially since XSS rules prevent us from
launching XMLHttpRequests to other domains (e.g. to grab the source
code).