Don't use goog.array.forEach in bookmarklet code (fixes issue 12).
This makes jsprettify work on Facebook. I have no idea _why_ this
makes jsprettify work on Facebook. It just does. goog.array is still
included (probably via goog.dom). The only difference---literally, only
difference, modulo some local variables pointing at different letters
in the alphabet---is that in this version, goog.array.forEach (which
corresponds to the local variable "m") is not defined, nor called, in
the output. That's _it_.

Here's, as far as I can tell, what was going on: If the compiled version
of goog.array.forEach (as the [local] variable "m") was called on the
output, then the jsprettify bookmarklet would cause some Facebook code
to throw "TypeError: Object #<a NodeList> has no method 'map'" when it
was run. At no point did jsprettify call "map" on a NodeList, nor is
it clear why the _local_ definition of "m" as goog.array.forEach would
cause Facebook to think that a NodeList contains a method "map", nor
is it clear why code included by Facebook would even be running, hence
throwing an error, if this condition held, but not if it didn't.

JavaScript is weird.